ml_service/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.45.0
pillow>=10.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
numpy>=1.24.0
scipy>=1.11.0
gunicorn>=21.2.0
